pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yourmemory"
version = "1.3.0"
description = "Persistent memory for Claude — Ebbinghaus forgetting curve, semantic deduplication, MCP-native"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Sachit Misra", email = "mishrasachit1@gmail.com" }
]
keywords = ["mcp", "claude", "memory", "ebbinghaus", "ai", "sqlite", "postgresql"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "python-dotenv",
    "mcp",
    "sentence-transformers",
    "numpy",
    "python-dateutil",
    "duckdb>=0.10.0",
    "apscheduler",
    "spacy>=3.8.13,<4.0",
    "networkx>=3.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary", "pgvector"]
sse = ["fastapi", "uvicorn[standard]", "httpx"]
neo4j = ["neo4j>=5.0"]
all = ["yourmemory[postgres,sse,neo4j]"]

[project.scripts]
yourmemory = "memory_mcp:run"
yourmemory-path = "memory_mcp:print_path"
yourmemory-setup = "memory_mcp:setup"

[project.urls]
Homepage = "https://github.com/sachitrafa/YourMemory"
Repository = "https://github.com/sachitrafa/YourMemory"

[tool.setuptools]
py-modules = ["memory_mcp"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["*.sql"]
